front: intervals editor, new segments have a unit by default (interval type number with unit)
